Break parsing code up into multiple modules
